+++ packages/zefyr/test/widgets/scope_test.dart
@@ -16,7 +16,6 @@ void main() {
       scope = ZefyrScope.editable(
         mode: ZefyrMode.edit,
         controller: ZefyrController(doc),
-        imageDelegate: ZefyrDefaultImageDelegate(),
         focusNode: FocusNode(),
         focusScope: FocusScopeNode(),
       );
@@ -27,7 +26,7 @@ void main() {
       scope.addListener(() {
         notified = true;
       });
-      final delegate = ZefyrDefaultImageDelegate();
+      final delegate = _TestImageDelegate();
       scope.imageDelegate = delegate;
       expect(notified, isTrue);
       notified = false;
@@ -73,4 +72,22 @@ void main() {
       expect(notified, isTrue);
     });
   });
+}
+
+class _TestImageDelegate implements ZefyrImageDelegate<String> {
+  @override
+  Widget buildImage(BuildContext context, String key) {
+    return null;
+  }
+
+  @override
+  String get cameraSource => null;
+
+  @override
+  String get gallerySource => null;
+
+  @override
+  Future<String> pickImage(String source) {
+    return null;
+  }
 }